fix: Change cmds to fix logging errors
While starting the container the log was filled with errors from
WeeWX, wich can´t use the log system. Starting rsyslog over init.d
fixed this.
Change startup cmd for weewx to run in daemon mode and retry forever
if the weatherstation is not available on startup
Clean up Dockerfile, insert a variable for python version